@@ -11,27 +11,27 @@
     "prettier": "prettier --write ."
   },
   "dependencies": {
-    "bootstrap": "^4.6.0",
+    "bootstrap": "^5.1.3",
     "classnames": "^2.3.1",
-    "localforage": "^1.9.0",
-    "react": "^17.0.0",
-    "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.0",
-    "react-router-dom": "^5.2.0"
+    "localforage": "^1.10.0",
+    "react": "^17.0.2",
+    "react-bootstrap": "^2.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.0.2"
   },
   "devDependencies": {
     "@rollup/plugin-eslint": "^8.0.1",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.7",
-    "@typescript-eslint/eslint-plugin": "^4.20.0",
-    "@typescript-eslint/parser": "^4.20.0",
-    "@vitejs/plugin-react-refresh": "^1.3.1",
-    "eslint-plugin-react": "^7.23.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "^2.2.1",
-    "typescript": "^4.1.2",
-    "vite": "^2.1.5"
+    "@types/react": "^17.0.35",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "eslint-plugin-react": "^7.27.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "prettier": "^2.4.1",
+    "typescript": "^4.5.2",
+    "vite": "^2.6.14"
   },
   "eslintConfig": {
     "parser": "@typescript-eslint/parser",
